refactor(luce-bench): decompose cli.py, dedupe helpers, wire CI + tests
Behavior-preserving cleanup of the luce-bench harness surfaced by review:

- Split the 1.9k-line cli.py into focused modules: _display.py (pure
  formatters, format_row decomposed into throughput/timing/tokens),
  _preflight.py (model resolution + server probing), _orchestrate.py
  (run/sweep orchestration). cli.py is now the arg-parse + dispatch shell.
- Dedupe: _resolve_auth_header (6 inline sites), _write_area_json /
  _terse_rows (7 envelope sites), schema._row_passed (3 pass-rule sites),
  and the two near-identical agent-replay drivers (_maybe_restart /
  _judge_or_pending).
- Fix datetime.utcnow() deprecation in probe.py (output unchanged).
- Add unit tests: _display, cli helpers, and probe.py (a shipped
  entry point that had zero coverage). New test pins the cost-containment
  invariant that agent_recorded never runs under --areas all.
- Wire the suite into CI (it previously never ran pytest) and add the
  luce-bench uv.lock that was omitted by oversight.

All public import paths preserved. 320 -> 358 tests, ruff clean.
